Reorganize textarea interface
