+++ app/core/interfaces/textarea/component.js
@@ -6,28 +6,15 @@
 //  For all details and documentation:
 //  http://www.getdirectus.com
 
-define(['app', 'core/UIComponent', 'core/UIView', 'core/t'],function(app, UIComponent, UIView, __t) {
+define([
+  'core/interfaces/textarea/interface',
+  'core/UIComponent',
+  'core/t'
+],function(Input, UIComponent, __t) {
 
   'use strict';
 
-  var template = '<textarea rows="{{rows}}" name="{{name}}" id="{{name}}" placeholder="{{placeholder_text}}" {{#if readonly}}readonly{{/if}}>{{value}}</textarea>';
-
-  var Input = UIView.extend({
-    templateSource: template,
-
-    serialize: function() {
-      return {
-        value: this.options.value,
-        name: this.options.name,
-        rows: this.options.settings.get('rows'),
-        placeholder_text: this.options.settings.get('placeholder_text'),
-        comment: this.options.schema.get('comment'),
-        readonly: !this.options.canWrite
-      };
-    }
-  });
-
-  var Component = UIComponent.extend({
+  return UIComponent.extend({
     id: 'textarea',
     dataTypes: ['TEXT', 'VARCHAR'],
     variables: [
@@ -47,6 +34,4 @@ define(['app', 'core/UIComponent', 'core/UIView', 'core/t'],function(app, UIComp
       return _.isString(options.value) ? options.value.replace(/<(?:.|\n)*?>/gm, '').substr(0,100) : '<span class="silver">--</span>';
     }
   });
-
-  return Component;
 });
