+++ app/core/interfaces/textarea/input.html
@@ -1,0 +1,1 @@
+<textarea rows="{{rows}}" name="{{name}}" id="{{name}}" placeholder="{{placeholder_text}}" {{#if readonly}}readonly{{/if}}>{{value}}</textarea>
